Moving test MapKmersTest to the correct new package and adding Ignores on the failing tests.
