+++ src/test/kotlin/net/maizegenetics/phgv2/pathing/MapKmersTest.kt
@@ -1,8 +1,7 @@
-package net.maizegenetics.phgv2.cli
+package net.maizegenetics.phgv2.pathing
 
 import com.github.ajalt.clikt.testing.test
-import net.maizegenetics.phgv2.pathing.MapKmers
-import net.maizegenetics.phgv2.pathing.ReadInputFile
+import net.maizegenetics.phgv2.cli.TestExtension
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
